requirements.txt
Flask
Flask-Babel
mediapipe
opencv-python
numpy
fpdf2